@Receipt.class Added the following methods: - whitespacePaddedProductQuantity(int quantity) - getProductsNamesValuesAndQuantity()
Fixed the toString() method and it now lists the products in three colums: Item, Price and Quantity
The length of a line in the Receipt file is 50 chars.

The Collection which the Receipt objects now use is a List of Sellable objects.
See below for details of the Sellable inteface.

@OutOfStockProductException.class

Created a new custom exception which is thrown when an item is out of stock
or when the cashier attempts to sell an item when there is no sufficient items
in storage.

@Sellable.java (Inteface)

The Sellable interface is used as a template for the SoldProduct object so that
all SoldProduct objects implement the necessary methods.